# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Grand Tour Sochi - Public API
  version: 1.2.0
  description: |-
    Публичный API для каталога и бронирования услуг Grand Tour Sochi.
    Ключевые возможности:
    - Получение списка маршрутов и доступных слотов.
    - Создание и управление бронированиями (draft -> hold -> invoice).
    - Идемпотентность для всех мутирующих запросов через заголовок `Idempotency-Key`.
...
paths:
...
  /api/booking/confirm:
    post:
      tags:
        - Booking
      summary: Подтвердить бронирование (оплата получена)
      description: Переводит бронирование из `invoice` в `confirm`. Вызывается оператором. Требует `Idempotency-Key`.
      operationId: confirmBooking
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookingId:
                  type: string
                  format: uuid
                paymentRef:
                  type: string
                  description: "Ссылка или номер подтверждения офлайн-оплаты."
      responses:
        "200":
          description: Бронирование успешно подтверждено.
          content:
            application/json:
              schema:
                type: object
                properties:
                  state:
                    type: string
                    example: "confirm"
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
          description: Человекочитаемое описание ошибки.
        code:
          type: string
          description: Машиночитаемый код ошибки.
        details:
          type: object
          additionalProperties: true
        traceId:
          type: string
          format: uuid
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: "Уникальный ключ для обеспечения идемпотентности запроса. Рекомендуется использовать UUID v4."
  responses:
    BadRequest:
      description: "Ошибка валидации или отсутствует обязательный заголовок, например `Idempotency-Key`."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
            properties:
              code:
                type: string
                example: "IDEMPOTENCY_KEY_REQUIRED"
    NotFound:
      description: "Запрашиваемый ресурс не найден."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: "Конфликт состояния. Например, попытка зарезервировать уже занятый слот."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...